Story.FromJson: Cleanup
Now that we're utilizing __init__, it's unecessary to manually set each
attribute for the Story. The method setattr() makes our job easier.

There are exceptions for requested_by_* due to the nature of the API
default response. Refer to the comments for more details.

Similarly, we handle created_at and deadline manually to parse datetime
into something more useful.