Story: Clean up requested_by_* attributes
Now that we have an implementation of ProjectMemberships, we'll look up
a Person through that path. There is a utility function under the
Tracker named getPersonById() which expects a ProjectMembership object
along with an ID.

This simplifies Story and prevents our dependency on modifying the API
call to always filter.